Extend database check
By starting the application it will be create the var-folder, so that the check will be false on every time and the script doesn't initialize the session tables. For fixing this the constructor takes a closer view if the sqlite-file exists. If not it knows that it is a new Sqlite-database.